+++ src/chaser/random_unicolor.rs
@@ -26,18 +26,26 @@ use super::{Chaser, TwoParameterChaser, UnicolorTransition};
 use crate::{sequence::Unicolor, time::TimeConfig};
 
 /// A LED chaser that performs random transitions.
-pub struct RandomUnicolor<D: Distribution<u32>, const N: usize> {
+pub struct RandomUnicolor<
+    HD: Distribution<i16>,
+    TD: Distribution<u32>,
+    const N: usize,
+> {
     /// The random number generator for color and transition speed selection.
     rng: SmallRng,
     /// The refresh rate.
     refresh_rate: Hertz,
+    /// The hue distribution.
+    hue_distr: HD,
     /// The transition speed distribution.
-    transition_time_distr: D,
+    transition_time_distr: TD,
     /// The ongoing transition.
     transition: UnicolorTransition<N>,
 }
 
-impl<const N: usize> Chaser<N> for RandomUnicolor<Uniform<u32>, N> {
+impl<const N: usize> Chaser<N>
+    for RandomUnicolor<Uniform<i16>, Uniform<u32>, N>
+{
     fn set_time_config(&mut self, time_config: &TimeConfig) {
         self.refresh_rate = time_config.refresh_rate;
 
@@ -56,8 +64,13 @@ impl<const N: usize> Chaser<N> for RandomUnicolor<Uniform<u32>, N> {
     }
 }
 
-impl<const N: usize> RandomUnicolor<Uniform<u32>, N> {
-    fn new(refresh_rate: Hertz, transition_time_distr: Uniform<u32>) -> Self {
+impl<const N: usize> RandomUnicolor<Uniform<i16>, Uniform<u32>, N> {
+    /// Builds a new random unicolor chaser.
+    pub fn new(
+        refresh_rate: Hertz,
+        hue_distr: Uniform<i16>,
+        transition_time_distr: Uniform<u32>,
+    ) -> Self {
         let mut rng = SmallRng::seed_from_u64(0);
 
         let start_color = Hsv {
@@ -69,6 +82,7 @@ impl<const N: usize> RandomUnicolor<Uniform<u32>, N> {
         let transition = generate_transition(
             &mut rng,
             refresh_rate,
+            &hue_distr,
             &transition_time_distr,
             start_color,
         );
@@ -76,13 +90,33 @@ impl<const N: usize> RandomUnicolor<Uniform<u32>, N> {
         Self {
             rng,
             refresh_rate,
+            hue_distr,
             transition_time_distr,
             transition,
         }
     }
+
+    /// Sets the color temperature.
+    ///
+    /// A negative values gives warmer hues, a positive one colder hues.
+    pub fn set_temperature(&mut self, temperature: i8) {
+        self.hue_distr = if temperature.is_negative() {
+            Uniform::new(
+                -128 + temperature.abs() as i16,
+                127 - temperature.abs() as i16,
+            )
+        } else {
+            Uniform::new(
+                temperature.abs() as i16,
+                255 - temperature.abs() as i16,
+            )
+        };
+    }
 }
 
-impl<D: Distribution<u32>, const N: usize> Iterator for RandomUnicolor<D, N> {
+impl<HD: Distribution<i16>, TD: Distribution<u32>, const N: usize> Iterator
+    for RandomUnicolor<HD, TD, N>
+{
     type Item = Unicolor<RGB8, N>;
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -92,6 +126,7 @@ impl<D: Distribution<u32>, const N: usize> Iterator for RandomUnicolor<D, N> {
                 self.transition = generate_transition(
                     &mut self.rng,
                     self.refresh_rate,
+                    &self.hue_distr,
                     &self.transition_time_distr,
                     self.transition.end_color().into(),
                 );
@@ -105,11 +140,12 @@ impl<D: Distribution<u32>, const N: usize> Iterator for RandomUnicolor<D, N> {
 fn generate_transition<const N: usize>(
     rng: &mut impl Rng,
     refresh_rate: Hertz,
+    hue_distr: &impl Distribution<i16>,
     transition_time_distr: &impl Distribution<u32>,
     start_color: Hsv,
 ) -> UnicolorTransition<N> {
     let end_color = Hsv {
-        hue: rng.gen(),
+        hue: rng.sample(hue_distr) as u8,
         sat: 255,
         val: 255,
     };